Move PaypalItemBuilder under Paypal module
